Probes/edge_storage/requirements.txt
prometheus-client
kubernetes
Flask
PyQt5
flask-httpauth
psutil
requests
pyyaml